# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Smart Investment Advisor",
    "description": "AI 智能投资助手后端 API，集成多源数据与 LLM 分析能力",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Read Users Me",
        "operationId": "read_users_me_api_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/user/password": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Change Password",
        "operationId": "change_password_api_user_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/user/settings": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update User Settings",
        "operationId": "update_user_settings_api_user_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/user/test-connection": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Test Ai Connection",
        "operationId": "test_ai_connection_api_user_test_connection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiConfig": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "type": "object",
        "title": "ApiConfig"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChange": {
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "PasswordChange"
      },
...
      "TestConnectionRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "TestConnectionRequest"
      },
      "TestConnectionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "TestConnectionResponse"
      },
...
      "UserProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "membership_tier": {
            "type": "string",
            "title": "Membership Tier"
          },
          "has_gemini_key": {
            "type": "boolean",
            "title": "Has Gemini Key"
          },
          "has_deepseek_key": {
            "type": "boolean",
            "title": "Has Deepseek Key"
          },
          "has_siliconflow_key": {
            "type": "boolean",
            "title": "Has Siliconflow Key"
          },
          "api_configs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ApiConfig"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Configs"
          },
          "fallback_enabled": {
            "type": "boolean",
            "title": "Fallback Enabled"
          },
          "preferred_data_source": {
            "type": "string",
            "title": "Preferred Data Source"
          },
          "preferred_ai_model": {
            "type": "string",
            "title": "Preferred Ai Model"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "feishu_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feishu Webhook Url"
          },
          "enable_price_alerts": {
            "type": "boolean",
            "title": "Enable Price Alerts"
          },
          "enable_hourly_summary": {
            "type": "boolean",
            "title": "Enable Hourly Summary"
          },
          "enable_daily_report": {
            "type": "boolean",
            "title": "Enable Daily Report"
          },
          "enable_macro_alerts": {
            "type": "boolean",
            "title": "Enable Macro Alerts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email",
          "membership_tier",
          "has_gemini_key",
          "has_deepseek_key",
          "has_siliconflow_key",
          "fallback_enabled",
          "preferred_data_source",
          "preferred_ai_model",
          "timezone",
          "theme",
          "enable_price_alerts",
          "enable_hourly_summary",
          "enable_daily_report",
          "enable_macro_alerts"
        ],
        "title": "UserProfile"
      },
      "UserSettingsUpdate": {
        "properties": {
          "api_key_gemini": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Gemini"
          },
          "api_key_deepseek": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Deepseek"
          },
          "api_key_siliconflow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Siliconflow"
          },
          "api_configs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ApiConfig"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Configs"
          },
          "fallback_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Enabled"
          },
          "preferred_data_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Data Source"
          },
          "preferred_ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Ai Model"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "feishu_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feishu Webhook Url"
          },
          "enable_price_alerts": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Price Alerts"
          },
          "enable_hourly_summary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Hourly Summary"
          },
          "enable_daily_report": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Daily Report"
          },
          "enable_macro_alerts": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Macro Alerts"
          }
        },
        "type": "object",
        "title": "UserSettingsUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}